pyproject.toml
[project]
name = "personal-assistant-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
    "ruff>=0.11.10",
    "pydantic>=2.11.7",
    "google-adk>=1.6.1",
    "pytz>=2025.2",
    "dateparser>=1.2.2",
]
